% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_model.R
\name{as_kv.model_system}
\alias{as_kv.model_system}
\title{Return KV rows for a model_system (system-level + all components)}
\usage{
\method{as_kv}{model_system}(x, ...)
}
\description{
Return KV rows for a model_system (system-level + all components)
}
\keyword{internal}
