% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_estimation_control.R
\name{define_estimation_control}
\alias{define_estimation_control}
\title{Define estimation control settings}
\usage{
define_estimation_control(num_cores = 1)
}
\arguments{
\item{num_cores}{Integer. Number of processes to use for parallel estimation (default = 1)}

\item{model_system}{Object of class model_system.}
}
\value{
An object of class estimation_control containing control settings
}
\description{
Define estimation control settings
}
