% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_factor_model.R
\name{define_factor_model}
\alias{define_factor_model}
\title{Define latent factor model structure}
\usage{
define_factor_model(
  n_factors,
  n_types,
  n_quad_points,
  correlation = FALSE,
  n_mixtures = 1
)
}
\arguments{
\item{n_factors}{Integer. Number of latent factors (>=1)}

\item{n_types}{Integer. Number of types (>=1)}

\item{n_quad_points}{Integer. Number of Gauss-Hermite quadrature points (>=1)}

\item{correlation}{Logical. Whether to allow correlation between two factors (default = FALSE)}

\item{n_mixtures}{Integer. Number of discrete mixtures (default = 1, allowed: 1-3)}
}
\value{
An object of class "factor_model"
}
\description{
Define latent factor model structure
}
