% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_model_component.R
\name{define_model_component}
\alias{define_model_component}
\title{Define a model component}
\usage{
define_model_component(
  name = NULL,
  data,
  outcome,
  factor,
  evaluation_indicator = NULL,
  covariates,
  model_type = c("linear", "logit", "probit"),
  intercept = TRUE,
  factor_normalization = 1,
  num_choices = 2,
  nrank = NULL
)
}
\arguments{
\item{name}{name of model component (Optional)}

\item{data}{data.frame for validation}

\item{outcome}{Character. Name of the outcome variable.}

\item{factor}{model. Object of Class factor_model.}

\item{evaluation_indicator}{Character (optional). Variable used for evaluation subsample.}

\item{covariates}{List of Character vectors. Names of covariates.}

\item{model_type}{Character. Type of model (e.g., "linear", "logit", "probit").}

\item{intercept}{Logical. Whether to include an intercept (default = TRUE).}

\item{factor_normalization}{Numeric. Factor loading normalization constant (default = 1).}

\item{num_choices}{Integer. Number of choices (for multinomial models).}

\item{nrank}{Integer (optional). Rank for exploded multinomial logit.}
}
\value{
An object of class "model_component". A list representing the model component
}
\description{
Define a model component
}
