% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_model_system.R
\name{define_model_system}
\alias{define_model_system}
\title{Define a model system}
\usage{
define_model_system(components, factor)
}
\arguments{
\item{components}{A named list of model_component objects}
}
\value{
An object of class "model_system". A list of model_component objects and one factor_model object.
}
\description{
Define a model system
}
