% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_model_component.R
\name{print.model_component}
\alias{print.model_component}
\title{Print method for model_component objects}
\usage{
\method{print}{model_component}(x, ...)
}
\arguments{
\item{x}{An object of class "model_component".}

\item{...}{Not used.}
}
\description{
Print method for model_component objects
}
